Fixed http-retry tests so they work when run with other tests...
The problem was the http-tests which were setting the env-var
CDB_DRY_RUN to "True" but not undoing this before the end
of the test, thus interfering with the next tests.

The tests are being run in simple alphabetical order.
We should look into running them in a random order to
help reveal test interdependencies like this.